requirements.txt
paho-mqtt>=1.3
pyserial>=3.2
pyzmq>=16.0.2
